requirements.txt
requests
python-frontmatter